+++ test/unit/Test.Unit.CLASP.Versions/CLASP_version_tester.cs
@@ -18,7 +18,7 @@ namespace Test.Unit.CLASP.Versions
             internal static class Expectations
             {
                 internal const int  VersionMajor    =   0;
-                internal const int  VersionMinor    =   17;
+                internal const int  VersionMinor    =   18;
             }
         }